requirements.txt
requests
beautifulsoup4
python-dotenv
faiss-cpu
numpy
rank-bm25
sentence-transformers
langchain
